+++ R/replace_mnirs.R
@@ -269,13 +269,25 @@ replace_invalid <- function(
     method <- match.arg(method)
 
     ## process ========================================================
-    y <- x
     ## fill invalid indices with NA
     invalid_idx <- c(
         which(x %in% invalid_values),
         which(x >= invalid_above),
         which(x <= invalid_below)
     )
+    invalid_length <- length(invalid_idx)
+
+    if (verbose) {
+        cli_inform(c(
+            "!" = "{.val {invalid_length}} invalid samples replaced.",
+        ))
+    }
+    
+    if (invalid_length == 0) {
+        return(x)
+    }
+    
+    y <- x
     y[invalid_idx] <- NA_real_
 
     if (method == "median") {
@@ -283,12 +295,8 @@ replace_invalid <- function(
             validate_width_span(width, span, verbose)
         }
 
-        window_idx <- compute_local_windows(
-            t, invalid_idx, width, span
-        )
-        local_medians <- compute_local_fun(
-            y, window_idx, median, na.rm = TRUE
-        )
+        window_idx <- compute_local_windows(t, invalid_idx, width, span)
+        local_medians <- compute_local_fun(y, window_idx, median, na.rm = TRUE)
         ## if method = "median"
         ## invalid_values removed to NA first,
         ## so returns local median excluding idx
@@ -357,15 +365,9 @@ replace_outliers <- function(
     method <- match.arg(method)
 
     ## process =====================================================
-    window_idx <- compute_local_windows(
-        t, width = width, span = span
-    )
-    local_medians <- compute_local_fun(
-        x, window_idx, median, na.rm = TRUE
-    )
-    is_outlier <- compute_outliers(
-        x, window_idx, local_medians, outlier_cutoff
-    )
+    window_idx <- compute_local_windows(t, width = width, span = span)
+    local_medians <- compute_local_fun(x, window_idx, median, na.rm = TRUE)
+    is_outlier <- compute_outliers(x, window_idx, local_medians, outlier_cutoff)
 
     ## fill outliers with median or NA
     y <- x